+++ CHANGES.md
@@ -9,6 +9,7 @@ New:
   to socket methods, added socket() method on srt input/outputs
   (#2556)
 - Added `video.align`.
+- Added `string.index`.
 
 Changed: